Arreglando limpiado de cache de las fotos del admin, YA QUE SEBA NO LO HIZO BIEEEN, AHHHH

+++ src/Loogares/BlogBundle/Entity/Posts.php
@@ -781,10 +781,10 @@ class Posts
         $this->upload();
 
         // Eliminamos thumbnails
-        if(file_exists(__DIR__.'/../../../../web/assets/media/cache/blog/'.$this->getUploadDir().'/'.$this->imagen))
+        if($this->imagen != '' && file_exists(__DIR__.'/../../../../web/assets/media/cache/blog/'.$this->getUploadDir().'/'.$this->imagen))
             unlink(__DIR__.'/../../../../web/assets/media/cache/blog/'.$this->getUploadDir().'/'.$this->imagen);
             
-        if(file_exists(__DIR__.'/../../../../web/assets/media/cache/blog_detalle/'.$this->getUploadDir().'/'.$this->imagen_detalle))
+        if($this->imagen_detalle != '' && file_exists(__DIR__.'/../../../../web/assets/media/cache/blog_detalle/'.$this->getUploadDir().'/'.$this->imagen_detalle))
             unlink(__DIR__.'/../../../../web/assets/media/cache/blog_detalle/'.$this->getUploadDir().'/'.$this->imagen_detalle);
     }
 }
